SOFT: Extract DES/3DES functions to mech_openssl.c
Extract the OpenSSL based DES/3DES functions into file mech_openssl.c,
so that they can be used by other tokens as well.
