+++ usr/lib/common/h_extern.h
@@ -2938,6 +2938,36 @@ CK_RV openssl_specific_aes_cmac(STDLL_TokData_t *tokdata, CK_BYTE *message,
                                 CK_ULONG message_len, OBJECT *key, CK_BYTE *mac,
                                 CK_BBOOL first, CK_BBOOL last, CK_VOID_PTR *ctx);
 
+CK_RV openssl_specific_des_ecb(STDLL_TokData_t *tokdata,
+                               CK_BYTE *in_data,
+                               CK_ULONG in_data_len,
+                               CK_BYTE *out_data,
+                               CK_ULONG *out_data_len,
+                               OBJECT *key, CK_BYTE encrypt);
+CK_RV openssl_specific_des_cbc(STDLL_TokData_t *tokdata,
+                               CK_BYTE *in_data,
+                               CK_ULONG in_data_len,
+                               CK_BYTE *out_data,
+                               CK_ULONG *out_data_len,
+                               OBJECT *key, CK_BYTE *init_v, CK_BYTE encrypt);
+CK_RV openssl_specific_tdes_ecb(STDLL_TokData_t *tokdata,
+                                CK_BYTE *in_data,
+                                CK_ULONG in_data_len,
+                                CK_BYTE *out_data,
+                                CK_ULONG *out_data_len,
+                                OBJECT *key, CK_BYTE encrypt);
+CK_RV openssl_specific_tdes_cbc(STDLL_TokData_t *tokdata,
+                                CK_BYTE *in_data,
+                                CK_ULONG in_data_len,
+                                CK_BYTE *out_data,
+                                CK_ULONG *out_data_len,
+                                OBJECT *key, CK_BYTE *init_v, CK_BYTE encrypt);
+CK_RV openssl_specific_tdes_mac(STDLL_TokData_t *tokdata, CK_BYTE *message,
+                                CK_ULONG message_len, OBJECT *key, CK_BYTE *mac);
+CK_RV openssl_specific_tdes_cmac(STDLL_TokData_t *tokdata, CK_BYTE *message,
+                                 CK_ULONG message_len, OBJECT *key, CK_BYTE *mac,
+                                 CK_BBOOL first, CK_BBOOL last, CK_VOID_PTR *ctx);
+
 #include "tok_spec_struct.h"
 extern token_spec_t token_specific;
